feat(ui): load larger history pages behind an in-flow earlier-history boundary
Back-scroll now requests 400-message pages (session open keeps its 100-message
tail) and arms the auto-load sentinel 1200px early, so older history mostly
lands before the reader hits the top. The floating 'Show earlier' pill becomes
an in-flow divider boundary above the oldest loaded message: idle it offers
Show earlier; while loading, an accent comet sweeps both hairlines into a
pulsing label. The boundary renders above the virtualized block and folds into
the virtualizer scrollMargin (composed with the thread's top padding), so
prepends keep re-anchoring on message rows; margin-only header toggles
compensate the scroll offset explicitly. Splits the private virtualizer host
out of chat-transcript-controller and deletes the dead page-up click branch.
